WIP allocating byte buffers via Netty
